requirements.txt
certifi==2020.12.5
chardet==4.0.0
Django==2.2.17
djangorestframework==3.12.2
filelock==3.0.12
idna==2.10
importlib-metadata==3.3.0
Markdown==3.3.3
pytz==2020.4
requests==2.25.1
requests-file==1.5.1
six==1.15.0
sqlparse==0.4.1
tldextract==3.1.0
typing-extensions==3.7.4.3
urllib3==1.26.2
zipp==3.4.0
